+++ compiler/TemplateBuilder.js
@@ -491,6 +491,11 @@ TemplateBuilder.prototype = {
         } else {
             params = ['out'];
         }
+
+        // Don't use "use strict" in compiled templates since it
+        // could break backwards compatibility.
+        // out.append('"use strict";\n');
+
         out.append('function create(__helpers) {\n');
         //Write out the static variables
         this.writer.flush();